+++ DMS/requirements.txt
@@ -1,0 +1,9 @@
+certifi==2023.11.17
+charset-normalizer==3.3.2
+idna==3.6
+numpy==1.26.2
+opencv-python==4.8.1.78
+openvino==2023.2.0
+openvino-telemetry==2023.2.1
+requests==2.31.0
+urllib3==2.1.0
